requirements.txt
absl-py==0.15.0
ansible==2.9.6
apache-libcloud==2.8.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
attrs==19.3.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.0.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.5
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debtcollector==2.5.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
eventlet==0.33.3
filelock==3.0.12
flatbuffers==1.12
fonttools==4.37.4
gast==0.3.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
greenlet==2.0.1
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.0.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==1.1.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.0
more-itertools==4.2.0
msgpack==1.0.4
mxnet==1.7.0.post2
netaddr==0.7.19
netifaces==0.10.4
networkx==2.8.8
nftables==0.1
ntlm-auth==1.1.0
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
onnx==1.13.0
opt-einsum==3.3.0
oslo.config==9.1.0
oslo.i18n==5.1.0
ovs==2.13.8
packaging==21.3
pandas==1.4.4
pbr==5.6.0
pexpect==4.6.0
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
pymongo==4.2.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRAPL==0.2.3.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.2.1
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
repoze.lru==0.7
requests==2.28.2
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==2.0.0
Routes==2.5.1
rsa==4.9
ryu==4.34
scikit-learn==1.1.2
scipy==1.9.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.15.0
sklearn==0.0
slip==0.6.5
slip.dbus==0.6.5
sortedcontainers==2.1.0
sqlparse==0.2.4
ssh-import-id==5.10
stevedore==3.3.0
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.4
tensorflow-estimator==2.4.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==3.1.0
tinyrpc==1.1.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
virtualenv==20.4.6
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
WebOb==1.8.7
websocket-client==0.53.0
Werkzeug==2.2.2
wrapt==1.12.1
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
